// src/common/dto/preApproved.dto.ts
export class PreApprovedDto {
    id: string;
    codCliente: string;
    codCuenta: string;
    promedioDeudas: string;
    porcentajeDeudas: string;
    montoPrestamoMax: string;
    edad: string;
    campana: string;
    antiguedad: string;
    califConti: string;
    califBCP: string;
    refAtraso: string;
    esCodeudor: string;
    mora: string;
    branch: string;
    listaNegativas: string;
    ros: string;
    cantPrestamos: string;
}